move testutil to cloudtest
